Add trashbox.ts
